+++ app/src/main/java/ca/etsmtl/applets/etsmobile/util/ETSMobileAuthenticator.java
@@ -13,8 +13,6 @@ import android.preference.PreferenceManager;
 import android.text.TextUtils;
 import android.util.Log;
 
-
-
 import org.json.JSONException;
 import org.json.JSONObject;
