+++ app/src/main/java/ca/etsmtl/applets/etsmobile/util/SignetsMethods.java
@@ -1,0 +1,31 @@
+package ca.etsmtl.applets.etsmobile.util;
+
+/**
+ * Method mapping for Signets-Mobile
+ *
+ * @author Phil
+ *
+ */
+public class SignetsMethods {
+    public static final int INFO_ETUDIANT = 0;
+    public static final int LIST_COURS = 3;
+    public static final int LIST_INT_SESSION = 4;
+    public static final int LIST_SESSION = 2;
+    public static final int LIST_PROGRAM = 6;
+    public static final int LIST_COEQ = 5;
+    public static final int LIST_EVAL = 1;
+    public static final int LIST_HORAIRE_PROF = 7;
+    public static final int LIRE_HORAIRE = 8;
+    public static final int LIRE_JOURS_REMPLACES = 9;
+    public static final int BOTTIN_LIST_DEPT = 10;
+    public static final int BOTTIN_GET_FICHE = 11;
+    public static final int BOTTIN_GET_FICHE_DATA = 12;
+    public static final int BOTTIN_GET_ALL = 13;
+    public static final int BOTTIN_GET_FICHE_BY_SERVICE = 14;
+    public static final int BOTTIN_GET_LIST_SERVICE_AND_EMP = 15;
+    public static final int LIST_EXAMENS_FINAUX = 16;
+    public static final int LIST_SEANCES = 17;
+    public static final int LIST_SEANCES_CURRENT_AND_NEXT_SESSION = 18;
+    public static final int LIST_EXAM_CURRENT_AND_NEXT_SESSION = 19;
+    public static final int LIST_JOURSREMPLACES_CURRENT_AND_NEXT_SESSION = 20;
+}
